move logic to accesscontroller to pass tests
