+++ src/Dorc.PersistentData/Sources/EnvironmentsPersistentSource.cs
@@ -953,112 +953,5 @@ namespace Dorc.PersistentData.Sources
                 context.SaveChanges();
             }
         }
-
-        public void UpdateAccessControlsWithHistory(
-            Guid objectId,
-            List<AccessControlApiModel> newPrivileges,
-            ClaimsPrincipal user)
-        {
-            using (var context = contextFactory.GetContext())
-            {
-                var environment = context.Environments.FirstOrDefault(e => e.ObjectId == objectId);
-                
-                if (environment == null)
-                    throw new ArgumentException("Environment not found");
-                    
-                var username = _claimsPrincipalReader.GetUserFullDomainName(user);
-                
-                // Get existing access controls
-                var existingAccessControls = context.AccessControls
-                    .Where(ac => ac.ObjectId == objectId)
-                    .ToList();
-                    
-                var existingDict = existingAccessControls.ToDictionary(ac => ac.Id, ac => ac);
-                var newIds = newPrivileges.Select(p => p.Id).ToArray();
-                
-                // Track removed users
-                foreach (var existing in existingAccessControls)
-                {
-                    if (!newIds.Contains(existing.Id))
-                    {
-                        // Create history for removal
-                        EnvironmentHistoryPersistentSource.AddHistory(
-                            environment,
-                            GetPermissionsString(existing.Allow),
-                            $"User '{existing.Name}' removed from access control",
-                            username,
-                            "Access Control - User Removed",
-                            context
-                        );
-                        
-                        // Remove the access control
-                        context.AccessControls.Remove(existing);
-                    }
-                }
-                
-                // Track added or modified users
-                foreach (var privilege in newPrivileges)
-                {
-                    if (privilege.Id == 0)
-                    {
-                        // NEW USER - Add to database
-                        var newAccessControl = new AccessControl
-                        {
-                            ObjectId = objectId,
-                            Name = privilege.Name,
-                            Sid = privilege.Sid,
-                            Pid = privilege.Pid,
-                            Allow = privilege.Allow,
-                            Deny = privilege.Deny
-                        };
-                        context.AccessControls.Add(newAccessControl);
-                        
-                        // Create history for addition
-                        EnvironmentHistoryPersistentSource.AddHistory(
-                            environment,
-                            string.Empty,
-                            $"User '{privilege.Name}' added with permissions: {GetPermissionsString(privilege.Allow)}",
-                            username,
-                            "Access Control - User Added",
-                            context
-                        );
-                    }
-                    else if (existingDict.TryGetValue(privilege.Id, out var existing))
-                    {
-                        // Check if permissions changed
-                        if (existing.Allow != privilege.Allow || existing.Deny != privilege.Deny)
-                        {
-                            // Create history for modification
-                            EnvironmentHistoryPersistentSource.AddHistory(
-                                environment,
-                                GetPermissionsString(existing.Allow),
-                                $"User '{privilege.Name}' permissions changed to: {GetPermissionsString(privilege.Allow)}",
-                                username,
-                                "Access Control - Permissions Modified",
-                                context
-                            );
-                            
-                            // Update the access control
-                            existing.Allow = privilege.Allow;
-                            existing.Deny = privilege.Deny;
-                            existing.Name = privilege.Name;
-                            existing.Sid = privilege.Sid;
-                            existing.Pid = privilege.Pid;
-                        }
-                    }
-                }
-                
-                context.SaveChanges();
-            }
-        }
-
-        private string GetPermissionsString(int permissions)
-        {
-            var perms = new List<string>();
-            if ((permissions & 1) != 0) perms.Add("Write");
-            if ((permissions & 2) != 0) perms.Add("Read Secrets");
-            if ((permissions & 4) != 0) perms.Add("Owner");
-            return perms.Any() ? string.Join(", ", perms) : "None";
-        }
     }
 }
